requirements.txt
Django==1.7.0
six==1.6.1  # import before WebTest
WebTest==2.0.15
python-dateutil==2.2
fabric==1.9.0
pep8==1.5.7
psycopg2==2.5.3
coverage==3.7.1
pylint==1.2.1
redis==2.10.1
celery==3.1.13
djangorestframework==2.3.14
djangorestframework-gis==0.2
markdown==2.4.1
django-bootstrap-form==3.1
django-filter==0.7
django-allauth==0.17.0
requests==2.4.0  # Manually added for djrill, do not remove
simplejson==3.5.2  # required for vectorformats
django_nose==1.2
newrelic==2.22.1.20
GPolyEncode==0.1.1
django-redis==3.7.1
honcho==0.5.0
kombu==3.0.22
boto==2.32.1
billiard==3.3.0.18
pytz==2014.7
sqlparse==0.1.11
statsd==3.0.1
django-push-notifications==1.1.0
djrill==1.1.1
